#285: split ChallengeController per parent resource
Pure refactor — 1193-line ChallengeController → three resource-per-
parent controllers (SignInChallengeController,
SignUpChallengeController, EmailAddressChallengeController) under
app/Http/Controllers/Fapi/. Shared helpers (createChallenge,
mapVerificationStatus, refreshChallengeFromVerification, etc.) move
to a ManagesChallenges trait. routes/fapi.php updated to point each
challenge route at its new controller; behavior unchanged.
